feat(memory): move memory tools to the memory library
